src/requirements.txt
sentry_sdk==0.14.3
google-cloud-bigquery
google-cloud-storage
